BugFixed

+++ Aban360.ReportPool.Domain/Features/BuiltIns/ServiceLinkTransaction/Outputs/FinancialStatementHeaderOutputDto.cs
@@ -9,18 +9,18 @@ namespace Aban360.ReportPool.Domain.Features.BuiltIns.ServiceLinkTransaction.Out
         public string Title { get; set; }
         public string ReportDateJalali { get; set; } = DateTime.Now.ToShortPersianDateString();
         public int RecordCount { get; set; }
-
-        public int CustomerCount { get; set; }
-        public int ConsumptionTotalUnit { get; set; }
-        public int DailyAverage { get; set; }
-        public int NetConsumption { get; set; }
+        
+        public long CustomerCount { get; set; }
+        public long ConsumptionTotalUnit { get; set; }
+        public decimal DailyAverage { get; set; }
+        public long NetConsumption { get; set; }
         public long NetAmount { get; set; }
-        public int ReturnedConsumption { get; set; }
+        public long ReturnedConsumption { get; set; }
         public long ReturnedAmount { get; set; }
         public long DiscountAmount { get; set; }
-        public int RawConsumption { get; set; }
+        public long RawConsumption { get; set; }
         public long RawAmount { get; set; }
-        public float RawAmountAverage { get; set; }
-        public float ConsumptionAverageInMonth { get; set; }
+        public decimal RawAmountAverage { get; set; }
+        public decimal ConsumptionAverageInMonth { get; set; }
     }
 }
